add additional notifications + fix window size + interval uprade

+++ src/views/ChessClaimView.py
@@ -81,7 +81,8 @@ class ChessClaimView(QMainWindow):
         """ Centers the window on the screen """
         screen = QDesktopWidget().screenGeometry()
         size = self.geometry()
-        self.move((screen.width() - size.width()) / 2, (screen.height() - size.height()) / 2)
+        self.move(int((screen.width() - size.width()) / 2),
+                  int((screen.height() - size.height()) / 2))
 
     def set_gui(self) -> None:
         """ Initialize GUI components. """
